Function vault::free_cstring

pub extern fn free_cstring(ptr: *mut c_char)

Extern function for deallocating a CString returned by parse_to_cstring.

Must only be passed a pointer created by parse_to_cstring; passing other pointers is undefined behaviour, and will likely cause a seg fault or double free. Every call to parse_to_cstring should have a matching free_cstring to deallocate the memory after the string has been used. Failure to call this function will result in a memory leak.