Function vault::parse_directory

pub fn parse_directory(path: &Path, config: Option<Config>) -> Result<Vec<Replay>>

Parses all replay and archive (.zip) files in the first level of the given directory. This function can be enabled with the parse-all feature.

If None is passed to config, the following default configuration is used:

strict = false
commands = true
command_bytes = false
clean_file = true

Examples

extern crate vault;

use std::path::Path;

let path = Path::new("/path/to/directory");
let replays = vault::parse_directory(&path, None).unwrap();