Function vault::parse_to_cstring

pub extern fn parse_to_cstring(path: *const c_char) -> *mut c_char

Extern function for invoking a parse operation across FFI. Returns a Vault type serialized to JSON.

Note that the return type is a pointer to a c_char array. The function passes ownership of the CString to the FFI caller and does not deallocate memory when the CString goes out of this function's scope. It is the responsibility of the FFI caller to pass back the *c_char to this library via free_cstring so that it can be deallocated. Failure to pass back to free_cstring will result in a memory leak.

Examples

// node.js

var ffi = require('ffi');
var ref = require('ref');

var charPtr = ref.refType(ref.types.CString);

var lib = ffi.Library('/path/to/vault/target/release/libvault', {
    'parse_to_cstring': [charPtr, ['string']],
    'free_cstring': ['void', [charPtr]]
});

var path = '/path/to/rec/zip/or/dir';
var ptr = lib.parse_to_cstring(path);
var str = ref.readCString(ptr, 0);
lib.free_cstring(ptr);

console.log(str);