Struct vault::Config

pub struct Config {
    pub strict: bool,
    pub commands: bool,
    pub command_bytes: bool,
}

This type contains option and configuration information for the associated Vault instance.

Fields

strict

commands

command_bytes

Methods

impl Config

fn new(strict: bool, commands: bool, command_bytes: bool) -> Config

Constructs a new Config structure with the given options set.