Struct vault::Item

pub struct Item {
    pub selection_id: u32,
    pub server_id: u32,
    pub item_type: ItemType,
}

This type contains a parsed representation of an item that can be equipped in a CoH2 replay.

Fields

selection_id

Used in a PCMD_SetCommander command to refer to the commander selected by a player

server_id

Corresponds to a server_id in Relic's attribute files

item_type

Internally used to organize different item types

Methods

impl Item

fn new(item_type: ItemType) -> Item

Constructs a new Item with an empty ID and the given ItemType.