Struct vault::Vault

pub struct Vault {
    pub replays: Vec<Replay>,
}

This type is the main entry point for the vault replay parser and provides the cleanest interface for use by external code.

Fields

replays

Methods

impl Vault

fn parse(path: &Path, strict: bool) -> Result<Vault>

Attempts to parse the given file, returning a Vault type populated with the Replay(s) if successful.

Currently .rec and .zip (archives) are supported filetypes. When an archive is provided, all .rec files in the archive will be parsed. All resulting Replays have their raw byte data cleaned automatically after parse completes.

Note that if the strict parameter is set to true, parsing with fail with an error if the path does not point to a directory or a file with a .rec or .zip extension.

Examples

extern crate vault;

use vault::Vault;
use std::path::Path;

let path = Path::new("/path/to/file");
let results = Vault::parse(&path, false).unwrap();

for replay in results.replays().iter() {
    println!("{}", replay.to_json());
}

fn to_json(&self) -> Result<String>

Serializes Vault as JSON String.