Crate vault

Vault is an in-memory database specification which is designed to work without a central server. The data in the database follows a versioned format where each user can access a chain of data that documents changes to a group or piece of related data over time. Only the user which holds the associated id and key may modify the data in a chain. Another owner can take control over the data if they know the id and the key.

Every Data chain starts with an InitTransaction. The InitTransaction contains the user's designated id, and some metadata. Any proceeding data on the same chain needs to be a descendent of this original record or else it is considered invalid. Data can be added to the chain via a DataTransaction. The DataTransaction is associated to the chain through the owner's ID and it contains its own randomly generated ID. As with every other record, a DataTransaction contains a Counter which allows the Vault to identify which record is the latest in the chain.

Records may also be revoked from the Vault through a RevocationTransaction. A RevocationTransaction is created and it references the id of a existing DataTransaction. The RevocationTransaction stages the associated record for deletion. The record is deleted when the chain preforms a garbage collection and the RevocationTransaction is deleted along with it.

Modules

base64

crypto_box

types

vault

Structs

ChainId
A chain identifier. Used to identify a set of transactions in a version.

DBReader
A reader for the DBView

DBView
A view over the records in a vault. Allows for loading and accessing the records.

DBWriter
A writer for the DBView. Allows mutation of the underlying data.

DeleteRequest
A delete a transaction Request. Contains the id and Kind of the transaction to be revoked.

Key
A key to the crypto box. Key is stored on the heap which makes it easier to erase.

ReadRequest
A read request call. Contains the ID and Kind for the referred transaction.

ReadResult
A Read Result which contains the data from the read request. Contains the Kind, id and data of the transaction. The data is guarded and needs to be unlocked and decrypted before it can be used.

RecordHint
a record hint. Used as a hint to what this data is.

RecordId
A record identifier. Contains a ChainID which refers to the "chain" of transactions in the Version.

WriteRequest
A Write Request. Contains the id, kind and data of the transaction that will be mutated.

Enums

Error
Errors for the Vault Crate

Kind
Enum that describes the type of a transaction.

PreparedRead
A preparation of the data to read from the database.

Statics

ALLOC
A Zeroing Allocator which wraps the standard memory allocator. This allocator zeroes out memory when it is dropped. Works on any application that imports stronghold.

Traits

Base64Decodable
a trait to make types base64 decodable

Base64Encodable
a trait to make types base64 encodable

BoxProvider
A provider interface between the vault and a crypto box. See libsodium's secretbox for an example.

Decrypt
Trait for decryptable data. Allows the data to be decrypted.

Encrypt
trait for encryptable data. Allows the data to be encrypted.

Type Definitions

Result