Module vault::service

Re-exports

pub use crate::connection_manager::warp;

pub use crate::metrics::monitor_bridge_metrics;

pub use crate::relay::Config;

pub use crate::relay::Runner;

Structs

CancellationScheduler

ConnectionManager

IssueCanceller

MonitoringConfig

ReplaceCanceller

ServiceConfig

ShutdownSender

A wrapper arround a tokio broadcast channel that makes sure that listeners created after a shutdown signal has already been sent also receive the shutdown signal.

Traits

Filter

Composable request filters.

Service

Functions

execute_open_requests

Queries the parachain for open requests and executes them. It checks the bitcoin blockchain to see if a payment has already been made.

init_subscriber

listen_for_accept_replace

Listen for AcceptReplaceEvent directed at this vault and continue the replacement procedure by transferring bitcoin and calling execute_replace

listen_for_execute_replace

Listen for ExecuteReplaceEvent directed at this vault and continue the replacement procedure by transferring bitcoin and calling execute_replace

listen_for_issue_cancels

Listen for all CancelIssueEvents.

listen_for_issue_executes

Listen for ExecuteIssueEvent directed at this vault. Cancels the scheduled cancel_issue

listen_for_issue_requests

Listen for RequestIssueEvent directed at this vault. Schedules a cancellation of the received issue

listen_for_redeem_requests

Listen for RequestRedeemEvent directed at this vault; upon reception, transfer bitcoin and call execute_redeem

listen_for_replace_requests

Listen for RequestReplaceEvent, and attempt to accept it

process_issue_requests

execute issue requests on best-effort (i.e. don’t retry on error), returns an error if stream ends, otherwise runs forever

spawn_cancelable

wait_or_shutdown

Type Definitions

DynBitcoinCoreApi