Struct vault::Item

pub struct Item {
    pub id: u64,
    pub item_type: ItemType,
}

This type contains a parsed representation of an item that can be equipped in a CoH2 replay.

Fields

id

item_type

Methods

impl Item

fn new(item_type: ItemType) -> Item

Constructs a new Item with an empty ID and the given ItemType.

fn with_split_id(primary: u32, secondary: u32, item_type: ItemType) -> Item

Constructs a new Item by combining the give u32 primary and secondary IDs into a single unique u64 ID, and passing through the given ItemType.

fn with_whole_id(id: u64, item_type: ItemType) -> Item

Constructs a new Item with the given ID and ItemType. This function simply passes through the given ID to the Item without any transformation.

fn update_id(&mut self, primary: u32, secondary: u32)

Combines u32 primary and secondary IDs into a single unique u64 ID by shifting the primary ID 32 bits to the left and then adding the two IDs together.